Create a helper XQueryCodePointRange class.
This class is designed to help write a Lexer that matches tokens
who's characters can match the full set of Unicode codepoints.

The current implementation is a simple Basic Multilingual Plane
matcher. This is to help define the interface needed by the Lexer
before extending the implementation to support the full set of
Unicode codepoints.