Implement PersistentStateComponent on IntelliJ <= 2018.2.
On IntelliJ >= 2018.3, a run configuration implementing
PersistentStateComponent results in serializeStateInto being
called instead of writeExternal. This causes the component
data to be missing due to an IntelliJ bug. Not implementing
PersistentStateComponent allows a compatibility fix for the
serialization bug to be used.

On IntelliJ <= 2018.2, RunCnfigurationBase.getOptionClass
does not use ConfigurationFactory.getOptionClass. Because
it is called during initialization, the base implementation
is always called. Implementing PersistentStateComponent
causes getStateClass to be called so the correct data class
is created.

This completes the compatibility fixes to make calls to
serializeConfigurationInto work on all supported IntelliJ
versions.